Revert to single-file widgets and move utils to module util folders.
