+++ assets/js/modules/analytics/components/dashboard/DashboardOverallPageMetricsWidget.js
@@ -1,7 +1,7 @@
 /**
- * DashboardOverallPageMetricsWidgetGA4 component.
+ * OverallPageMetricsWidget component.
  *
- * Site Kit by Google, Copyright 2023 Google LLC
+ * Site Kit by Google, Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,26 +29,25 @@ import Data from 'googlesitekit-data';
 import {
 	DATE_RANGE_OFFSET,
 	MODULES_ANALYTICS,
-} from '../../../datastore/constants';
-import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
-import { CORE_SITE } from '../../../../../googlesitekit/datastore/site/constants';
-import { MODULES_ANALYTICS_4 } from '../../../../analytics-4/datastore/constants';
-import { Grid, Row, Cell } from '../../../../../material-components/layout';
-import PreviewBlock from '../../../../../components/PreviewBlock';
-import DataBlock from '../../../../../components/DataBlock';
-import Sparkline from '../../../../../components/Sparkline';
-import SourceLink from '../../../../../components/SourceLink';
-import whenActive from '../../../../../util/when-active';
-import { generateDateRangeArgs } from '../../../util/report-date-range-args';
-import { getURLPath } from '../../../../../util';
-import WidgetHeaderTitle from '../../../../../googlesitekit/widgets/components/WidgetHeaderTitle';
-import useViewOnly from '../../../../../hooks/useViewOnly';
-import { calculateOverallPageMetricsData } from './utils';
+} from '../../datastore/constants';
+import { CORE_USER } from '../../../../googlesitekit/datastore/user/constants';
+import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import { Grid, Row, Cell } from '../../../../material-components/layout';
+import PreviewBlock from '../../../../components/PreviewBlock';
+import DataBlock from '../../../../components/DataBlock';
+import Sparkline from '../../../../components/Sparkline';
+import SourceLink from '../../../../components/SourceLink';
+import whenActive from '../../../../util/when-active';
+import { generateDateRangeArgs } from '../../util/report-date-range-args';
+import { calculateOverallPageMetricsData } from '../../util/overall-page-metrics';
+import { getURLPath } from '../../../../util';
+import WidgetHeaderTitle from '../../../../googlesitekit/widgets/components/WidgetHeaderTitle';
+import useViewOnly from '../../../../hooks/useViewOnly';
 const { useSelect, useInViewSelect } = Data;
 
-function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
+function DashboardOverallPageMetricsWidget( { Widget, WidgetReportError } ) {
 	const isGatheringData = useInViewSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).isGatheringData()
+		select( MODULES_ANALYTICS ).isGatheringData()
 	);
 
 	const viewOnlyDashboard = useViewOnly();
@@ -66,26 +65,23 @@ function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 
 	const args = {
 		...dates,
-		dimensions: [ 'date' ],
+		dimensions: [ 'ga:date' ],
 		metrics: [
 			{
-				name: 'screenPageViews',
+				expression: 'ga:pageviews',
+				alias: 'Pageviews',
 			},
 			{
-				name: 'sessions',
+				expression: 'ga:uniquePageviews',
+				alias: 'Unique Pageviews',
 			},
 			{
-				name: 'engagedSessions',
+				expression: 'ga:bounceRate',
+				alias: 'Bounce Rate',
 			},
 			{
-				name: 'averageSessionDuration',
-			},
-		],
-		orderby: [
-			{
-				dimension: {
-					dimensionName: 'date',
-				},
+				expression: 'ga:avgSessionDuration',
+				alias: 'Session Duration',
 			},
 		],
 		url,
@@ -102,16 +98,13 @@ function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 
 	const isLoading = useSelect(
 		( select ) =>
-			! select( MODULES_ANALYTICS_4 ).hasFinishedResolution(
-				'getReport',
-				[ args ]
-			)
+			! select( MODULES_ANALYTICS ).hasFinishedResolution( 'getReport', [
+				args,
+			] )
 	);
 
 	const error = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getErrorForSelector( 'getReport', [
-			args,
-		] )
+		select( MODULES_ANALYTICS ).getErrorForSelector( 'getReport', [ args ] )
 	);
 
 	const serviceURL = useSelect( ( select ) => {
@@ -126,7 +119,7 @@ function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 	} );
 
 	const report = useInViewSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getReport( args )
+		select( MODULES_ANALYTICS ).getReport( args )
 	);
 
 	const currentDayCount = useSelect( ( select ) =>
@@ -173,7 +166,7 @@ function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 		);
 	}
 
-	const data = calculateOverallPageMetricsData( report, dates.startDate );
+	const data = calculateOverallPageMetricsData( report );
 
 	return (
 		<Widget Header={ Header } Footer={ Footer }>
@@ -214,5 +207,5 @@ function DashboardOverallPageMetricsWidgetGA4( { Widget, WidgetReportError } ) {
 }
 
 export default whenActive( { moduleName: 'analytics' } )(
-	DashboardOverallPageMetricsWidgetGA4
+	DashboardOverallPageMetricsWidget
 );
